requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
streamlit>=1.32.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
pytest>=7.4.0
